Add files via upload
Create an AnimalThread class that extends the Thread class.
The AnimalThread class must have fields for the thread name and thread priority. In the run() method, set up a counter for the meters that the animals have covered.
Organize the creation and launch of threads in the RabbitAndTurtle class. Use the constructor to pass the stream name and priority.