Add files via upload
You have $15. In the store, you will find out the price for a chocolate bar — $ 1.
The seller also informs you that he will give you another chocolate bar for 3 wrappers. What is the maximum amount of chocolates сan you expect?
Write a program for variable input data, the task of which will be to find the maximum of chocolates:
money: Available money
price: Price per chocolate bar
wrap: The number of wrappers needed to get another chocolate bar.
We can assume that all input data are integers and greater than 1.