Add files via upload
Create an Account class to represent a bank account. The Account class should contain the balance field to place the balance of the bank account, as well as methods for replenishing the balance and withdrawing money from the account.

Organize multiple replenishment of the balance in a separate stream for an arbitrary amount (can be set randomly).

In the Account class, implement a method that allows you to wait for the balance to be replenished to the required amount for withdrawing money from the account.

The main class should:

launch a stream to top up your account,
call a method that allows you to withdraw money as soon as the specified amount is accumulated,
and withdraw the balance on the balance.
Ensure thread-safe operation of the application.